Update the condition in deleteUser controller for checking the user is permitted or not to delete the user

+++ src/controllers/user.controller.ts
@@ -95,19 +95,19 @@ export const getAllUsers = async (req: Request, res: Response, next: NextFunctio
 };
 
 // Controller method for deleting a user
-export const deleteUser = async (req: RequestWithUser, res: Response, next: NextFunction) => {
+export const deleteUser = async (req: Request, res: Response, next: NextFunction) => {
     try {
         // Extract the userId from the request parameters
         const userId = req.params.userId;
 
         // Extract the logged-in user's ID from the request
-        // const loggedInUserId = req.user._id;
+        const loggedInUserId = (req.user as UserDocument)._id;
 
-        // // Check if the userId from the request matches the logged-in user's ID
-        // if (userId !== loggedInUserId.toString() ) {
-        //     // If not, return an error response indicating that only the user can delete their own account
-        //     return res.status(403).json({ success: false, error: 'You can only delete your own account.' });
-        // }
+        // Check if the userId from the request matches the logged-in user's ID
+        if (userId !== loggedInUserId.toString() || (req.user as UserDocument).role !== 'admin') {
+            // If not, return an error response indicating that only the user can delete their own account
+            return res.status(403).json({ success: false, error: 'You are not able to delete.' });
+        }
 
         // Attempt to find and delete the user by their ID
         const result = await User.findByIdAndDelete(userId).exec();
